commit2

+++ tema3.practica3/TareasWeb/src/main/java/org/example/tareasweb/repositories/EquipoRepository.java
@@ -3,8 +3,8 @@ package org.example.tareasweb.repositories;
 import org.example.tareasweb.entities.Equipo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
-public interface EquipoRepository extends JpaRepository<Equipo, Integer> {
+public interface EquipoRepository extends JpaRepository<Equipo, Long> {
+
 }
