+++ tema3.practica3/TareasWeb/src/main/java/org/example/tareasweb/repositories/TrabajadorRepository.java
@@ -3,8 +3,8 @@ package org.example.tareasweb.repositories;
 import org.example.tareasweb.entities.Trabajador;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
-public interface TrabajadorRepository extends JpaRepository<Trabajador, Integer> {
+public interface TrabajadorRepository extends JpaRepository<Trabajador, Long> {
+
 }
